SCC: renames reader file
