+++ src/main/python/ttconv/scc/reader.py
